@@ -3,18 +3,28 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@emotion/core": "^10.0.22",
-    "@material-ui/core": "^3.9.4",
-    "@material-ui/icons": "^3.0.2",
-    "http-proxy-middleware": "^0.19.1",
-    "react": "^16.13.0",
+    "@emotion/core": "^11.0.0",
+    "@emotion/react": "^11.8.1",
+    "@emotion/styled": "^11.8.1",
+    "@mui/icons-material": "^5.4.4",
+    "@mui/material": "^5.4.4",
+    "@mui/styles": "^5.4.4",
+    "@testing-library/jest-dom": "^5.16.2",
+    "@testing-library/react": "^12.1.3",
+    "@testing-library/user-event": "^13.5.0",
+    "3d-force-graph": "^1.70.9",
+    "color": "^4.2.1",
+    "http-proxy-middleware": "^2.0.3",
+    "prop-types": "^15.8.1",
+    "query-string": "^7.1.1",
+    "react": "^17.0.2",
     "react-countup": "^6.1.1",
-    "react-dom": "^16.8.6",
+    "react-dom": "^17.0.2",
     "react-json-editor-ajrm": "^2.5.13",
-    "react-router-dom": "^5.0.0",
-    "react-scripts": "^3.2.0",
-    "react-spinners": "^0.5.4",
-    "typescript": "^3.4.5"
+    "react-router-dom": "^6.2.2",
+    "react-scripts": "^5.0.0",
+    "react-spinners": "^0.11.0",
+    "web-vitals": "^2.1.4"
   },
   "scripts": {
     "start": "react-scripts start",
