@@ -1,10 +1,10 @@
 import React, {Component} from "react";
 import PropTypes from 'prop-types';
-import Typography from '@material-ui/core/Typography';
-import {withStyles} from '@material-ui/core/styles';
-import Paper from '@material-ui/core/Paper';
+import Typography from '@mui/material/Typography';
+import {withStyles} from '@mui/material/styles';
+import Paper from '@mui/material/Paper';
 import {Link} from "react-router-dom";
-import FormattedContent from "./formattedContent";
+import FormattedContent from "./FormattedContent";
 
 const styles = theme => ({
   container: {
