@@ -1,5 +1,5 @@
-const proxy = require('http-proxy-middleware');
+const { createProxyMiddleware } = require('http-proxy-middleware');
 
 module.exports = function(app) {
-  app.use(proxy('/images', { target: process.env.REACT_APP_SERVER_URL }));
+  app.use(createProxyMiddleware('/images', { target: process.env.REACT_APP_SERVER_URL , changeOrigin:true }));
 };
